requirements.txt
brotlipy==0.7.0
cycler==0.11.0
graphviz==0.20.1
image-similarity-measures==0.3.6
imageio==2.28.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.3.3
mkl-fft==1.3.1
mkl-service==2.4.0
networkx==3.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
phasepack==1.5
Pillow==9.3.0
pyFFTW==0.13.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-ignite==0.4.12
pytz==2023.3
PyWavelets==1.4.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==1.13.1
torchaudio==0.13.1
torchsummary==1.5.1
torchvision==0.14.1
torchviz==0.0.2
tqdm==4.65.0
tzdata==2023.3

